+++ test/helpers/strapi.js
@@ -1,10 +1,7 @@
 'use strict';
 
 const _ = require('lodash');
-const fs = require('fs');
-const path = require('path');
 const dotenv = require('dotenv');
-const Knex = require('knex');
 const strapi = require('../../packages/core/strapi/lib');
 const { createUtils } = require('./utils');
 
@@ -16,94 +13,7 @@ const superAdminCredentials = {
 };
 
 const superAdminLoginInfo = _.pick(superAdminCredentials, ['email', 'password']);
-
-const databases = {
-  postgres: {
-    client: 'postgres',
-    connection: {
-      connectionString: 'postgresql://strapi:strapi@127.0.0.1',
-    },
-  },
-  mysql: {
-    client: 'mysql',
-    connection: {
-      host: '127.0.0.1',
-      port: 3306,
-      database: 'strapi_test',
-      username: 'strapi',
-      password: 'strapi',
-    },
-  },
-  sqlite: {
-    client: 'sqlite',
-    connection: {
-      filename: './tmp/data.db',
-    },
-    useNullAsDefault: true,
-  },
-};
-
-function CreateDB(uuid) {
-  // You can dynamically pass the database name
-  // as a command-line argument, or obtain it from
-  // a .env file
-  async function main(uuid) {
-    const knex = Knex(databases[process.env.DATABASE]);
-
-    // Lets create our database if it does not exist
-    await knex.raw(`CREATE DATABASE ${uuid};`);
-    process.env.DATABASE_NAME = uuid;
-    // knex.destory();
-  }
-  if (process.env.DATABASE !== 'sqlite') {
-    main(uuid).catch(console.log);
-  }
-}
-
-async function copyFolderRecursiveSync(source, target, noSkipBaseName) {
-  let files = [];
-
-  // check if folder needs to be created or integrated
-  let targetFolder = path.join(target, path.basename(source));
-  if (noSkipBaseName !== true) {
-    targetFolder = target;
-  }
-  try {
-    await fs.promises.access(targetFolder);
-  } catch (error) {
-    await fs.promises.mkdir(targetFolder);
-  }
-
-  // copy
-  const sourcePath = await fs.promises.lstat(source);
-  if (sourcePath.isDirectory()) {
-    files = await fs.promises.readdir(source);
-    for (const file of files) {
-      const curSource = path.join(source, file);
-      const curSourcePath = await fs.promises.lstat(curSource);
-      if (curSourcePath.isDirectory()) {
-        await copyFolderRecursiveSync(curSource, targetFolder, true);
-      } else {
-        await fs.promises.copyFile(curSource, path.join(targetFolder, file));
-      }
-    }
-  }
-}
-const runners = [];
-const createStrapiLoader = async () => {
-  if (runners.includes(process.env.JEST_WORKER_ID)) {
-    return;
-  }
-  runners.push(process.env.JEST_WORKER_ID);
-  const uuid = `strapi_${process.env.JEST_WORKER_ID.replace(/-/g, '')}`;
-  await copyFolderRecursiveSync(`./testApps/testApp`, `./testApps/${uuid}`);
-  CreateDB(uuid);
-  // await generateTestApp({ appName: uuid, database: sqlite, changeDBFile: true });
-  // read .env file as it could have been updated
-  dotenv.config({ path: process.env.ENV_PATH });
-  process.env.PORT = 0;
-  process.env.UUID = uuid;
-};
+const createStrapiLoader = async () => {};
 
 const createStrapiInstance = async ({
   ensureSuperAdmin = true,
@@ -113,8 +23,8 @@ const createStrapiInstance = async ({
   // read .env file as it could have been updated
   dotenv.config({ path: process.env.ENV_PATH });
   const options = {
-    appDir: `./testApps/${process.env.UUID}`,
-    distDir: `./testApps/${process.env.UUID}`,
+    appDir: `./testApps/strapi_${process.env.JEST_WORKER_ID}`,
+    distDir: `./testApps/strapi_${process.env.JEST_WORKER_ID}`,
   };
   const instance = strapi(options);
